Dictionary Generator
This program turns user input into a machine-readable dictionary, and two human-readable lists. One list shows which items each character likes, and the other shows which characters like each item. Both lists are written to the same .txt file, making it easy to search with ctrl + f.  It also allows the user to name the file so that the previous file isn't overwritten each time the program is run. This also makes it possible to quickly generate lists of gift preferences for multiple Harvest Moon games, or any other application that requires lists of compatibilies.